feat: create a file for duration interface
