+++ src/date/duration.ts
@@ -1,0 +1,9 @@
+export interface Duration {
+  day?: number
+  month?: number
+  year?: number
+  hour?: number
+  minutes?: number
+  seconds?: number
+  milliseconds?: number
+}
