+++ src/index.ts
@@ -5,3 +5,4 @@ export * from './lambda';
 export * from './number';
 export * from './object';
 export * from './string';
+export * from './date'
